feat: regexes now match _parts_ of filename
The match pattern regex does not have to match the entire path now. You
can for example do

powermv "(\d)" "{{_1|pad(2)}}" *

and it will replace the first occurance of a numerical digit in every
file (and directory) with a 2 digit padded version.

If you _want_ to match the entire path, then just ad ^ and $.